Assistant/services/agents/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "drm-chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-storage-blob>=12.28.0",
    "azure-identity>=1.15.0",
    "azure-mgmt-storage>=24.0.0",
    "boto3>=1.42.13",
    "botocore>=1.42.13",
    "click>=8.3.0",
    "duckdb>=1.4.0",
    "google-auth>=2.27.0",
    "hvac>=2.0.0",
    "logfire[starlette]>=4.9.0",
    "openai>=2.6.0",
    "pandas>=2.3.2",
    "pydantic-ai-slim[ag-ui,bedrock,google,mcp,openai,retries]>=1.39.0",
    "psycopg[binary]>=3.1.12",
    "pyodbc>=5.2.0",
    "pywin32>=306; sys_platform == 'win32'",
    "sqlalchemy>=2.0.43",
    "tabulate>=0.9.0",
    "uvicorn>=0.37.0",
    "opentelemetry-api>=1.28.0",
    "rich>=14.0.0",
    "azure-architecture-catalog-builder @ file://../../",
]
[tool.uv]
package = true
cache-keys = [{ file = "pyproject.toml" }, { git = { commit = true } }]

[tool.uv-dynamic-versioning]
vcs = "git"
style = "pep440"
bump = false

[project.scripts]
agents = "agents.server:main"

[tool.hatch.build.targets.wheel]
packages = ["agents"]

[tool.logfire]
project_name = "drm-chat"
send_to_logfire = "if-token-present"
